+++ Bot/Cogs/kumiko/gws.py
@@ -13,7 +13,8 @@ from genshin_wish_sim_utils import (
     KumikoWSUsersUtils,
     KumikoWSUtils,
 )
-from rin_exceptions import ItemNotFound, NoItemsError
+from kumiko_ui_components import GWSDeleteOneInvView, GWSPurgeAllInvView
+from rin_exceptions import NoItemsError
 
 load_dotenv()
 
@@ -29,73 +30,25 @@ wsUserInvUtils = KumikoWSUserInvUtils()
 wsUserUtils = KumikoWSUsersUtils()
 
 
-class PurgeView(discord.ui.View):
-    async def on_timeout(self):
-        for child in self.children:
-            child.disabled = True
-
-    @discord.ui.button(
-        label="Yes",
-        row=0,
-        style=discord.ButtonStyle.primary,
-        emoji=discord.PartialEmoji.from_str("<:check:314349398811475968>"),
-    )
-    async def button_callback(self, button, interaction):
-        checkUserInv = await wsUserInvUtils.getUserInv(
-            user_id=interaction.user.id, uri=WS_CONNECTION_URI
-        )
-        try:
-            if len(checkUserInv) == 0:
-                raise NoItemsError
-            else:
-                await wsUserInvUtils.purgeUserInv(
-                    user_id=interaction.user.id, uri=WS_CONNECTION_URI
-                )
-                await interaction.response.send_message(
-                    "Everything has been purged from your inventory. This cannot be recovered from.",
-                    ephemeral=True,
-                )
-        except NoItemsError:
-            await interaction.response.send_message(
-                "It seems like you don't have anything in your GWS inventory. Please try again",
-                ephemeral=True,
-            )
-
-    @discord.ui.button(
-        label="No",
-        row=0,
-        style=discord.ButtonStyle.primary,
-        emoji=discord.PartialEmoji.from_str("<:xmark:314349398824058880>"),
-    )
-    async def second_button_callback(self, button, interaction):
-        await interaction.response.send_message(
-            "Well glad you choose not to...", ephemeral=True
-        )
-
-
 class GWS(commands.Cog):
     """Kumiko's Genshin Wish Simulator (GWS)"""
 
     def __init__(self, bot):
         self.bot = bot
 
-    gws = SlashCommandGroup(
-        "gws", "Commands for Kumiko's Genshin Wish Sim", guild_ids=[970159505390325842]
-    )
-    gwsWish = gws.create_subgroup(
-        "wish", "Wish for some items", guild_ids=[970159505390325842]
-    )
+    gws = SlashCommandGroup("gws", "Commands for Kumiko's Genshin Wish Sim")
+    gwsWish = gws.create_subgroup("wish", "Wish for some items")
 
-    gwsUserInvDelete = gws.create_subgroup(
-        "delete", "Deletes some stuff from your inv", guild_ids=[970159505390325842]
-    )
+    gwsUserInvDelete = gws.create_subgroup("delete", "Deletes some stuff from your inv")
 
     @gwsWish.command(name="one")
-    async def gwsWishOne(self, ctx):
+    async def gwsWishOne(self, ctx: discord.ApplicationContext):
         """Allows you to wish for one item"""
+        await ctx.defer()
         getUserProfile = await wsUserUtils.getUserProfile(
             user_id=ctx.author.id, uri=WS_CONNECTION_URI
         )
+        starRank = 3
         for itemsMain in getUserProfile:
             if int(dict(itemsMain)["pulls"]) % 10 == 0:
                 starRank = 4
@@ -150,7 +103,7 @@ class GWS(commands.Cog):
                 await wsUserInvUtils.insertWSItemToUserInv(
                     uuid=mainResDict["uuid"],
                     user_id=ctx.author.id,
-                    date_obtained=datetime.now().isoformat(),
+                    date_obtained=discord.utils.utcnow().isoformat(),
                     name=mainResDict["name"],
                     description=mainResDict["description"],
                     star_rank=mainResDict["star_rank"],
@@ -166,23 +119,25 @@ class GWS(commands.Cog):
         embed.set_image(
             url=f"https://raw.githubusercontent.com/No767/Kumiko-WS-Assets/master/assets/{mainResDict['uuid']}.png"
         )
-        await ctx.respond(embed=embed)
+        await ctx.send_followup(embed=embed)
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
     @gwsWish.command(name="multiple")
     async def gwsWishMultiple(
         self,
-        ctx,
+        ctx: discord.ApplicationContext,
         *,
         num_of_wishes: Option(
-            int, "The number of wishes to perform", min_value=1, max_value=3
+            int, "The number of wishes to perform", min_value=1, max_value=10
         ),
     ):
         """Allows you to wish for up to 3 items at once"""
+        await ctx.defer()
         getUserProfile = await wsUserUtils.getUserProfile(
             user_id=ctx.author.id, uri=WS_CONNECTION_URI
         )
+        starRank = 3
         for itemsMain in getUserProfile:
             if int(dict(itemsMain)["pulls"]) % 10 == 0:
                 starRank = 4
@@ -237,7 +192,7 @@ class GWS(commands.Cog):
                     await wsUserInvUtils.insertWSItemToUserInv(
                         uuid=dict(mainItems)["uuid"],
                         user_id=ctx.author.id,
-                        date_obtained=datetime.now().isoformat(),
+                        date_obtained=discord.utils.utcnow().isoformat(),
                         name=dict(mainItems)["name"],
                         description=dict(mainItems)["description"],
                         star_rank=dict(mainItems)["star_rank"],
@@ -245,7 +200,6 @@ class GWS(commands.Cog):
                         amount=1,
                         uri=WS_CONNECTION_URI,
                     )
-
         mainPages = pages.Paginator(
             pages=[
                 discord.Embed(
@@ -314,37 +268,22 @@ class GWS(commands.Cog):
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
     @gwsUserInvDelete.command(name="one")
-    async def deleteOneUserInv(
-        self, ctx, *, name: Option(str, "The name of the item to delete")
-    ):
+    async def deleteOneUserInv(self, ctx):
         """Deletes one item from your inventory"""
-        checkUserInv = await wsUserInvUtils.getWSItemUserInvOne(
-            user_id=ctx.author.id, name=name, uri=WS_CONNECTION_URI
+        embed = discord.Embed()
+        embed.description = "Are you sure you want to delete this item?"
+        await ctx.respond(
+            embed=embed, view=GWSDeleteOneInvView(uri=WS_CONNECTION_URI), ephemeral=True
         )
-        try:
-            if len(checkUserInv) == 0:
-                raise ItemNotFound
-            else:
-                for items in checkUserInv:
-                    await wsUserInvUtils.deleteOneUserInv(
-                        user_id=ctx.author.id,
-                        uuid=dict(items)["item_uuid"],
-                        uri=WS_CONNECTION_URI,
-                    )
-                    await ctx.respond(
-                        f"The item {name} has been deleted from your inventory"
-                    )
-        except ItemNotFound:
-            embedError = discord.Embed()
-            embedError.description = "Sorry, but it seems like there are no items in your inventory. Please try again"
-            await ctx.respond(embed=embedError)
 
     @gwsUserInvDelete.command(name="all")
     async def purgeUserInv(self, ctx):
         """Deletes all of your items in your GWS inventory"""
         embed = discord.Embed()
         embed.description = "Do you want to purge all of the items from your user inventory? This cannot be undone"
-        await ctx.respond(embed=embed, view=PurgeView())
+        await ctx.respond(
+            embed=embed, view=GWSPurgeAllInvView(uri=WS_CONNECTION_URI), ephemeral=True
+        )
 
     @gws.command(name="profile")
     async def getUserProfile(self, ctx):
@@ -365,7 +304,7 @@ class GWS(commands.Cog):
                 inline=True,
             )
             embed.set_thumbnail(url=getUserInfo.display_avatar.url)
-        await ctx.respond(embed=embed)
+        await ctx.respond(embed=embed, ephemeral=True)
 
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())